+++ R/recode.R
@@ -15,7 +15,8 @@
 #' @param by A names vector (`new = old`); any non-matching values are set to
 #'   the appropriate `NA`
 #' @param vals An optional vector of values to use in lieu of a names in the
-#'   vector; this takes priority over `names(by)`
+#'   vector; this takes priority over `names(by)`.  This can be the same length
+#'   as `by` or a single value.
 #' @param mode passed to `as.vector()`
 #' @return A vector of values from `x`
 #'
@@ -35,8 +36,13 @@
 recode_by <- function(x, by, vals = NULL, mode = "any") {
   vals <- vals %||% names(by)
 
-  if (is.null(vals))
+  if (is.null(vals)) {
     stop("values to recode by were not properly set", call. = FALSE)
+  }
+
+  if (length(vals) == 1) {
+    vals <- rep.int(vals, length(by))
+  }
 
   as.vector(vals[match(x, by)], mode = mode)
 }
@@ -46,12 +52,20 @@ recode_by <- function(x, by, vals = NULL, mode = "any") {
 recode_only <- function(x, by, vals = NULL) {
   vals <- vals %||% names(by)
 
-  if (is.null(vals))
+  if (is.null(vals)) {
     stop("values to recode by were not properly set", call. = FALSE)
+  }
 
-  m <- match(x, by, nomatch = 0)
+  vals <- as.vector(vals, if (typeof(x) == "integer") "integer" else mode(x))
+
+  if (length(vals) == 1L) {
+    x[x %in% by] <- vals
+    return(x)
+  }
+
+  m <- match(x, by, nomatch = 0L)
   mode <- mode(x)
-  x[m > 0] <- vals[m]
+  x[m > 0L] <- vals[m]
   clean_na_coercion(as.vector(x, mode = mode))
 }