+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 
 * `unique.fact()` S3 method [#86](https://github.com/jmbarbone/mark/issues/86)
+* `recode_only()` and `recode_by()` can accept a single value for `val` [#72](https://github.com/jmbarbone/mark/issues/72)
 
 ## Fixes